### 2022-08-18 #554 @santikka
dynamite: Bayesian Modeling and Causal Inference for Multivariate Longitudinal Data
Submitting Author Name: Santtu Tikka
Submitting Author Github Handle: @santikka
Other Package Authors Github handles: @helske
Repository:  https://github.com/santikka/dynamite
Version submitted: 0.0.1
Submission type: Stats
Badge grade: gold
Version accepted: TBD
---

```
Package: dynamite
Title: Bayesian Modeling and Causal Inference for Multivariate
    Longitudinal Data
Version: 0.0.1
    person("Santtu", "Tikka", , "santtuth@gmail.com", role = c("aut", "cre"),
    person("Jouni", "Helske", , "jouni.helske@iki.fi", role = "aut",
           comment = c(ORCID = "0000-0001-7130-793X"))
  )
Description: Easy-to-use and efficient interface for
  Bayesian inference of complex panel (time series) data. The package supports
  joint modeling of multiple measurements per individual, time-varying and
  time-invariant effects, and a wide range of discrete and
  continuous distributions. Estimation of the models is carried out via 'Stan'.
    bayesplot,
    checkmate,
    cli,
    data.table (>= 1.14.3),
    glue,
    MASS,
    posterior,
    rstan (>= 2.26.11),
    utils
    covr,
    plm,
    knitr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd",
    "srr::srr_stats_roclet"))
LazyDataCompression: xz

```

- [x] A pre-submission inquiry has been approved in [issue #552](https://github.com/ropensci/software-review/issues/552)

The package is mainly intended for applied researchers working with complex panel data.
Panel data is common in many scientific fields, especially in sociology and econometrics. For example, analysing individual-level life-course data is valuable for assessing the effects of policy reforms and other interventions.

The `dynamite` R package provides easy-to-use interface for Bayesian inference of complex panel (time series) data comprising of multiple measurements per multiple individuals measured in time. The main features distinguishing the package and the underlying methodology from many other approaches are:

* Support for both time-invariant and time-varying effects modeled via B-splines.
* Joint modeling of multiple measurements per individual (multiple channels) based directly on the assumed data generating process.
* Support for non-Gaussian observations: Currently Gaussian, Categorical, Poisson, Bernoulli, Binomial, Negative Binomial, Gamma, Exponential, and Beta distributions are available and these can be mixed arbitrarily in multichannel models.
* Allows evaluating realistic long-term counterfactual predictions which take into account the dynamic structure of the model by posterior predictive distribution simulation.
* Transparent quantification of parameter and predictive uncertainty due to a fully Bayesian approach.
* User-friendly and efficient R interface with state-of-the-art estimation via [Stan](https://mc-stan.org/users/interfaces/rstan).

There are several R packages in CRAN focusing on panel data analysis including but not limited to:

* [`plm`](https://cran.r-project.org/web/packages/plm/index.html) for linear panel data models.
* [`fixest`](https://cran.r-project.org/web/packages/fixest/index.html) for fixed effects and different distributions of response variables (based on `stats::family`).
* [`panelr`](https://cran.r-project.org/web/packages/panelr/index.html) for "within-between" models which combine fixed effect and random effect models.
* [`lavaan`](https://cran.r-project.org/web/packages/lavaan/index.html) for general structural equation modelling, and thus can be used to estimate various panel data models such as cross-lagged panel models with fixed or random intercepts.

However, to the best of our knowledge, there are no other R packages (or software in general) that support all features of `dynamite` simultaneously. Thus *The first implementation of a novel algorithm* seems most applicable.

Gold.

We designed `dynamite` from the ground up with the standards in mind and thus we think the package fulfills all four aspects. We have been able to comply with 129 standards with only 42 N/A standards across the "Bayesian and Monte Carlo" and "Regression and Supervised Learning" categories. The package is very general and capable of supporting a wide range of data and model structures as demonstrated by the examples and the package tests. In our opinion, the internal structure of the package is also well motivated and compartmentalized. We have also tried to carefully select the package dependencies and keep them to a minimum.

  `autotest` reports failures for the following tests:
  ```r
      type       test_name        fn_name          parameter parameter_type   operation                      content
     <chr>      <chr>            <chr>            <chr>     <chr>            <chr>                          <chr>
   1 diagnostic single_char_case get_code         time      single character upper-case character parameter is case dependent
   2 diagnostic single_char_case get_code         group     single character upper-case character parameter is case dependent
   3 diagnostic single_char_case get_data         time      single character upper-case character parameter is case dependent
   4 diagnostic single_char_case get_data         group     single character upper-case character parameter is case dependent
   5 diagnostic single_char_case get_priors       time      single character upper-case character parameter is case dependent
   6 diagnostic single_char_case get_priors       group     single character upper-case character parameter is case dependent
   7 message    NA               mcmc_diagnostics NA        NA               normal function call           0 of 400 iterati…
   8 message    NA               mcmc_diagnostics NA        NA               normal function call           0 of 400 iterati…
   9 message    NA               mcmc_diagnostics NA        NA               normal function call           E-BFMI indicated…
  10 message    NA               mcmc_diagnostics NA        NA               NA                             0 of 400 iterati…
  11 message    NA               mcmc_diagnostics NA        NA               NA                             0 of 400 iterati…
  12 message    NA               mcmc_diagnostics NA        NA               NA                             E-BFMI indicated…
  ```
  The first six of these report that the `time` and `group` parameters of several functions are case dependent. This is by design, as both of these parameters correspond to column names of the input data. While not entirely sure, we believe the remaining failures occur because the function `mcmc_diagnostic` produces messages while not being a print or a summary method, but this is also by design as the function outputs diagnostic messages related to the MCMC runs. The output of `autotest` is not very clear on this matter, as all columns are just filled with `NA` values.
