### 2022-08-18 santikka/dynamite#46 @mpadge
Add "Remotes" field to DESC
@santikka Your submission to rOpenSci (https://github.com/ropensci/software-review/issues/554) failed checks because you've got minimal versions for both [`rstan`](https://cran.r-project.org/web/packages/rstan/index.html) and [`data.table`](https://cran.r-project.org/web/packages/data.table/index.html) that are beyond current CRAN versions (respectively 2.21.5 and 1.14.2, as of 18 Aug 2022). Your package fails to install because these version requirements can not be met.

You need to add these extra lines to your DESCRIPTION file:
```
Remotes:
  stan-dev/rstan,
  Rdatatable/data.table
```
That then tells the `r-lib` suite of installers (devtools, remotes) where to find those more recent versions, and so enables them to be installed. See docs of [the `remotes` package](https://github.com/r-lib/remotes) for more info. Thanks :smile:
