### 2022-08-18 santikka/dynamite#46 @santikka
@mpadge Using the Remotes field doesn't seem to work for rstan, because the github repo doesn't actually contain the package. For some reason, I also got a compilation error when trying to install data.table from Rdatatable/data.table. Adding this to the DESCRIPTION file seems to work, at least when installing locally:
```r
Additional_repositories: https://mc-stan.org/r-packages/,
    https://Rdatatable.gitlab.io/data.table
```
