### 2022-08-19 santikka/dynamite#46 @mpadge
@santikka Indeed, the "Remotes" field fails, and this "Additional_repositories" does work, but it only installs the latest release of the package, which for "rstan" is the current CRAN version. Their suggested method to ["install the latest development version of RStan"](https://github.com/stan-dev/rstan/wiki/RStan-Getting-Started) then does not work, and there is no systematic way to get any system to install your minimal required version of >= 2.26.11. I would suggest you either:

1. need to take this issue up with the Stan dev team, and ask them to actually release a version that satisfies that, whether on CRAN or via their https://mc-stan.org/r-packages/ repository server. Without that, nobody else will really be able to install `dynamite` either.
2. Revise `dynamite` to somehow relax that depenency back to current CRAN / mc-stan.org version(s).

I've also updated our system so the bot would issue a notification that:
> Note: The following R packages were unable to be installed/upgraded on our system: [rstan, StanHeaders]; some checks may be unreliable

I'm happy to keep tweaking our system to help here, so feel free to ask any further questions about how we can best solve this issue.
