### 2022-08-19 santikka/dynamite#46 @santikka
@mpadge Are you sure? If run this locally it installs `rstan` and `StanHeaders` version 2.26.13:
```r
remove.packages(c("rstan", "StanHeaders"))
remotes::install_github("santikka/dynamite")
```
Also, our gh workflows for codecov and R CMD check manage to install the required versions by adding the repositories under `extra-repositories` in the yaml files. Could this be included in your system as well? For example our `test-coverage.yaml` is:
```yaml
on:
  push:
    branches: [main, master]
  pull_request:
    branches: [main, master]

name: test-coverage

jobs:
  test-coverage:
    runs-on: ubuntu-latest
    env:
      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
      DYNAMITE_EXTENDED_TESTS: ${{contains(github.event.head_commit.message, 'run-extended')}}

    steps:
      - uses: actions/checkout@v2

      - uses: r-lib/actions/setup-r@v2
        with:
          extra-repositories: |
            https://mc-stan.org/r-packages/
            https://Rdatatable.gitlab.io/data.table
          use-public-rspm: true

      - uses: r-lib/actions/setup-r-dependencies@v2
        with:
          cache-version: 2
          extra-packages: |
            covr
            BH
            RcppParallel
            RcppEigen
            Rcpp
      - name: Test coverage
        run: covr::codecov()
        shell: Rscript {0}
````
