### 2022-08-19 santikka/dynamite#46 @mpadge
> Are you sure?

Unfortuantely, yes, but since it works for you, it then suggests that it comes down to idiosyncracies of our build system. We use a non-standard docker container which enables full installation of any required system libraries that are then able to be cross-linked on compilation of any other packages. This present issue obviously reveals one disadvantage of our system. If it's okay, can you please keep this issue open while I investigate further to see what can be done. I'll get back on to it first thing next week. thanks :smile:

(The gh-workflows file thing is something we discussed, but there really is no way to parse and use that information systematically, short of an assumption that everybody uses entirely standard `r-lib`-type workflows, which would be overly restrictive.)
