### 2022-08-19 santikka/dynamite#46 @santikka
That is unfortunate, I'll keep the issue open for now.

I'll also try to see if there is some way to use the Remotes field in DESC to install rstan and StanHeaders instead of the additional repos, if that is something that your system already supports.
