### 2022-08-20 santikka/dynamite#46 @mpadge
Thanks for helping with this @santikka . The problem definitely lies on our side now. I've got one idea on how i might be able to fix it, and will implement that this coming week. I'll close this issue on your repo now, and will let you know elsewhere when it's fixed.
