### 2022-08-24 #554 @santikka
@mpadge About the return values: the system does not seem to recognize function aliases. For example in the above, `obs` and `aux` are aliases of `dynamiteformula` and thus have the same return value implicitly. If I explicitly add the return tag to these aliases as well, the resulting documentation looks rather silly with regards to the output value:
```
Value
A dynamiteformula object.

A dynamiteformula object.

A dynamiteformula object.
```
