### 2022-09-21 santikka/dynamite#48 @helske
get_* functions do not work without group argument
[@nicholasjclark](https://github.com/ropensci/software-review/issues/554) noted during the review that one of the extended tests is currently broken as no grouping variable is used in the `get_priors` call. This should default to `NULL` as in `dynamite`.
