### 2022-09-21 #554 @noamross
Thank you for your review, @nicholasjclark!  I note you used the non-statistical reviewer template.  Would you be able to use the statistical one at https://stats-devguide.ropensci.org/pkgreview.html#pkgrev-template?  My apologies, it looks like the one linked above is dead so you probably searched out and found that one.

I am still seeking a second reviewer, @santikka. I suggest limiting updates until you have inputs from both.
