### 2022-09-28 #554 @santikka
Hi @nicholasjclark,

Thank your for the comprehensive review. We agree that the documentation in its current form is still lacking, and as we move toward writing the relevant papers, the situation will gradually improve as a consequence.

We have corrected the error in the unit testing. We have also corrected the errors and typos in the vignette.

While the native pipe is not strictly necessary, we still prefer it over the `magrittr` pipe because it is future-proof in a sense and is already widely adopted such as in RStudio and in the newest edition of the popular Advanced R book. The native pipe also has zero overhead in performance.

We have managed to drop the dependency on the development versions of `rstan` and `StanHeaders`. This only affects the case where the model contains categorical responses, for which the sampling will be much slower with older versions of these packages. A warning is now issued in such instances.

We have added the option to use `cmdstanr` as a backend for sampling which can be used with argument `backend = "cmdstanr"` in the `dynamite` call.

We have added helpful annotations to the more complicated section of the generated `Stan` code as well as in other parts of the package code itself. Some larger functions such as `dynamite` itself have been broken down to smaller simpler functions.

We have added an example on how to simulate from the prior predictive distribution to the documentation pages of `predict`. Note that this is not particularly useful for the complex models of `dynamite` because as with other flexible time-series models which do not enforce stationarity, the simulated series tend to vary too much and can "explode" with large number of time points.
