### 2022-12-06 #554 @LucyMcGowan
---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a gold badge
- [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

For packages aiming for silver or gold badges:

- [ ] This package extends beyond minimal compliance with standards in the following ways: (please describe)

---

### General Review

- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case

#### Algorithms

The algorithms are encoded well. The language is appropriate, and all tests passed on my machine.

This package has extensive testing and all tests passed on my machine.

The examples and vignette, however, are not currently compiling, for example, when I run the first example of the `dynamite` function I get the following error:

``` r
library(dynamite)
fit <- dynamite(
       dformula = obs(y ~ -1 + varying(~x), family = "gaussian") +
         lags(type = "varying") +
         splines(df = 20), gaussian_example, "id", "time",
       chains = 1,
       refresh = 0
     )
#> Error in `[.data.table`(data, idx, cl, env = list(cl = cl)): unused argument (env = list(cl = cl))
```

#### Visualisation (where appropriate)

- Do visualisations aid the primary purposes of statistical interpretation of results?
I am able to see the visualization in the README, they look good. Unfortunately, I am not able to compile the vignette due to the error above.

- Are there any aspects of visualisations which could risk statistical misinterpretation?
No

#### Package Design

- Is the package well designed for its intended purpose?
Yes
- In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
Yes
- In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
Yes
- In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
Yes
- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?
The output seems to work well with common functions, like those from the tidyverse.

---
