### 2022-12-07 #554 @santikka
@LucyMcGowan Thank you for the review!

Regarding the error, `dynamite` currently requires the development version of the `data.table` package (which can be installed via `data.table::update.dev.pkg()`). There is some confusion about the version numbers, since the required version `1.14.3` of `data.table` is already on CRAN, but the particular feature used by dynamite was not included in the CRAN release: https://github.com/Rdatatable/data.table/issues/5538 we will update the readme accordingly to avoid confusion.
