### 2022-12-08 #554 @nicholasjclark
## Package Review

-   Briefly describe any working relationship you may have (had) with the package authors (or otherwise remove this statement)

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

### Compliance with Standards

-   ☒ This package complies with a sufficient number of standards for a (bronze/silver/gold) badge
-   ☒ This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

For packages aiming for silver or gold badges:

-   ☒ This package extends beyond minimal compliance with standards in the following ways: (please describe)

The authors have been able to comply with 129 standards with only 42 N/A standards across the “Bayesian and Monte Carlo” and “Regression and Supervised Learning” categories and the function design is very well compartmentalised in my view

### General Review

#### Documentation

-   ☒ **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
-   ☒ The documentation is sufficient to enable general use of the package beyond one specific use case

The following sections of this template include questions intended to be used as guides to provide general, descriptive responses. Please remove this, and any subsequent lines that are not relevant or necessary for your final review.

#### Algorithms

-   How well are algorithms encoded?
-   Is the choice of computer language appropriate for that algorithm, and/or envisioned use of package?
-   Are aspects of algorithmic scaling sufficiently documented and tested?
-   Are there any aspects of algorithmic implementation which could be improved?

As the maintainers have now allowed for `cmdstanr` to be used as a backend, any efficiency updates can generally be automatically incorporated as long as users keep their `Cmdstan` and `cmdstanr` packages up to date. Perhaps some instructions to the users in the `Readme` can help to clarify this

#### Testing

-   Regardless of actual coverage of tests, are there any fundamental software operations which are not sufficiently expressed in tests?
-   Is there a need for extended tests, or if extended tests exists, have they been implemented in an appropriate way, and are they appropriately documented?

Very impressive and thorough set of unit tests

#### Visualisation (where appropriate)

-   Do visualisations aid the primary purposes of statistical interpretation of results?
-   Are there any aspects of visualisations which could risk statistical misinterpretation?

Visualisations make use of `ggplot` objects, which are very familiar to most potential users. There is no risk of misinterpretation in my view

#### Package Design

-   Is the package well designed for its intended purpose?
-   In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?
-   In relation to **External Design:** Do exported functions best serve inter-operability with other packages?
-   In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?
-   In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?

Function designs and levels of documentation are excellent and the returned structures will integrate very well with popular post-processing packages designed to work with `stanfit` objects. The algorithms make use of `stan`, which is the most efficient implementation of MCMC in the `R` programming environment. I see no issues with the way the package is designed

-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

Estimated hours spent reviewing: 5

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.
