### 2022-12-08 #554 @LucyMcGowan
Thank you! I updated to the dev version of `data.table` and was able to compile everything. I did try to follow the instructions in the README by running `data.table::update.dev.pkg()` but got the following message: `Error: 'update.dev.pkg' is not an exported object from 'namespace:data.table'`. I have updated my review above. Thank you!
