### 2022-12-08 #554 @santikka
@LucyMcGowan It seems that at some point the development version installation function of `data.table` was renamed into `update_dev_pkg` instead, apologies! The README has been corrected accordingly.
