### 2022-12-09 #554 @santikka
Based on the reviews, we have taken some additional steps to make the package installation more straightforward and the requirements less strict:

- The package no longer depends on `dplyr` or `tidyr`, and the internal package code that previously used these packages is now written entirely using `data.table` instead. `dplyr` and `tidyr` are only used in examples and tests, and are now included as 'Suggests' instead.
- The package no longer depends on R version 4.1.0. The native pipe `|>` is no longer used internally, and is used only in the examples conditionally on the installed R version of the user.
- The categorical distribution is now also supported for older `rstan` and `cmdstanr` versions.
