### 2022-12-09 #554 @noamross
Thank you all for bearing with us and following up! @LucyMcGowan and @nicholasjclark, you seem to have indicated in your follow-ups, but to be unambiguous, do the updates resolve any outstanding comments for acceptance at gold level?
