### 2022-12-09 #554 @LucyMcGowan
Yes, I approve acceptance at the gold level. Thanks @noamross!
