### 2022-12-13 #554 @nicholasjclark
> Thank you all for bearing with us and following up! @LucyMcGowan and @nicholasjclark, you seem to have indicated in your follow-ups, but to be unambiguous, do the updates resolve any outstanding issues for acceptance at gold level?

Yes I approve acceptance as well. Thanks very much
